cpp-port batch 104: port 29 hw/char/ + hw/misc/ files to C++, fix xhci switch + aes.h
Ported files:
- hw/char/: avr_usart, exynos4210_uart, max78000_uart, mchp_pfsoc_mmuart,
  renesas_sci, sh_serial, shakti_uart, sifive_uart, xilinx_uartlite (9 files)
- hw/misc/: armsse-cpu-pwrctrl, axp2xx, bcm2835_mphi, bcm2835_rng,
  exynos4210_rng, imx7_snvs, imx7_src, imx_rngc, max78000_aes,
  max78000_gcr, max78000_icc, max78000_trng, mps2-fpgaio, mps2-scc,
  msf2-sysreg, nrf51_rng, sifive_u_otp, slavio_misc, tz-mpc, tz-ppc (20 files)

Fixes:
- hcd-xhci.cpp: add default cases to TRB_TYPE switches (enum completeness)
- include/crypto/aes.h: add extern "C" guards for C++ compatibility
- Fix casts and extern "C" in existing .cpp files

+++ hw/char/cadence_uart.cpp
@@ -341,7 +341,7 @@ static void uart_write_tx_fifo(CadenceUARTState *s, const uint8_t *buf,
         return;
     }
 
-    if (size > CADENCE_UART_TX_FIFO_SIZE - s->tx_count) {
+    if (size > static_cast<int>(CADENCE_UART_TX_FIFO_SIZE - s->tx_count)) {
         size = CADENCE_UART_TX_FIFO_SIZE - s->tx_count;
         /*
          * This can only be a guest error via a bad tx fifo register push,
